add components

+++ modal/user.js
@@ -34,12 +34,18 @@ function insertUser(req,res){
 }
 
 function deleteUser(req,res){
+
+    if (req.body.id === undefined) {
+        console.log(req.body.id)
+        res.status(404).json({ ERROR : 'Not Found' });
+        return;
+    }
+
     const id = req.body.id;
     const insert = "DELETE FROM user WHERE user_id=?";
 
     connection.query(insert, id, function (err, result) {
         if (err) res.json(err.code);
-        console.log("Number of records inserted !: " + result.affectedRows);
         return res.json({state:'200'});
     });
 }
